// Angular/project/src/app/modelos/crear-oferta-inf-eco.interface.ts
export interface infoEconoI {
  id_oferta: Number | null | undefined;
  valor_oferta_inicial: Number | null | undefined;
  porcentaje_negociacion: Number | null | undefined;
  valor_oferta_final: Number | null | undefined;
  valor_terreno: Number | null | undefined;
  valor_construccion_m2: Number | null | undefined;
  valor_area_privada: Number | null | undefined;
  valor_cultivo: Number | null | undefined;
  avaluo_catastral: Number | null | undefined;
  valor_administracion: Number | null | undefined;
  valor_arriendo_inicial: Number | null | undefined;
  valor_arriendo_final: Number | null | undefined;
  valor_terraza_balcon_patio: Number | null | undefined;
  valor_garajes: Number | null | undefined;
  valor_depositos: Number | null | undefined;
  valor_anexidades: Number | null | undefined;
}